feat: refactor serialize & clean cargo.toml
