+++ manta-trusted-setup/Cargo.toml
@@ -32,12 +32,10 @@ rayon = ["dep:rayon", "manta-util/rayon"]
 std = []
 
 [dependencies]
-ark-bls12-381 = {version = "0.3.0", default-features = false, features = ["curve", "scalar_field"]} # TODO: Move to [dependencies-dev]
 ark-ec = { version = "0.3.0", default-features = false }
 ark-ff = { version = "0.3.0", default-features = false }
 ark-poly = {version = "0.3.0", default-features = false}
 ark-groth16 = { version = "0.3.0", default-features = false }
-ark-r1cs-std = { version = "0.3.1", default-features = false }
 ark-relations = { version = "0.3.0", default-features = false }
 ark-serialize = { version = "0.3.0", default-features = false, features = ["derive"] }
 ark-snark = { version = "0.3.0", default-features = false }
@@ -45,7 +43,6 @@ ark-std = { version = "0.3.0", default-features = false }
 blake2 = { version = "0.10.4", default-features = false, features = ["std"] }
 byteorder = { version = "1.4.3", default-features = false, features = ["std"] }
 derivative = { version = "2.2.0", default-features = false, features = ["use_core"] }
-manta-accounting = { path = "../manta-accounting", default-features = false }
 manta-crypto = { path = "../manta-crypto", default-features = false, features = ["getrandom"] }
 manta-util = { path = "../manta-util", default-features = false, features = ["alloc"] }
 rand = { version = "0.8.5", default-features = false, features = ["getrandom"] }
@@ -54,5 +51,7 @@ rayon = { version = "1.5.3", optional = true, default-features = false }
 workspace-hack = { version = "0.1.0", path = "../workspace-hack" }
 
 [dev-dependencies]
+ark-bls12-381 = {version = "0.3.0", default-features = false, features = ["curve", "scalar_field"]}
+ark-r1cs-std = { version = "0.3.1", default-features = false }
 ark-serialize = { version = "0.3.0", default-features = false, features = ["std"] }
-manta-pay = { path = "../manta-pay", default-features = false, features = ["download", "groth16", "scale", "scale-std", "std", "test"] } # TODO: To be removed
+manta-pay = { path = "../manta-pay", default-features = false, features = ["groth16"] } # TODO: To be removed
