+++ manta-trusted-setup/src/lib.rs
@@ -26,4 +26,3 @@ extern crate alloc;
 pub mod groth16;
 pub mod mpc;
 pub mod util;
-pub mod serialize;
